+++ q_shared.c
@@ -69,12 +69,12 @@ void RotatePointAroundVector(vec3_t dst, const vec3_t dir, const vec3_t point, f
 	im[2][1] = m[1][2];
 
 	memset(zrot, 0, sizeof(zrot));
-	zrot[0][0] = zrot[1][1] = zrot[2][2] = 1.0F;
+	zrot[0][0] = zrot[1][1] = zrot[2][2] = 1.0f;
 
-	zrot[0][0] = cos(DEG2RAD(degrees));
-	zrot[0][1] = sin(DEG2RAD(degrees));
-	zrot[1][0] = -sin(DEG2RAD(degrees));
-	zrot[1][1] = cos(DEG2RAD(degrees));
+	zrot[0][0] = cosf(DEG2RAD(degrees));
+	zrot[0][1] = sinf(DEG2RAD(degrees));
+	zrot[1][0] = -sinf(DEG2RAD(degrees));
+	zrot[1][1] = cosf(DEG2RAD(degrees));
 
 	R_ConcatRotations(m, zrot, tmpmat);
 	R_ConcatRotations(tmpmat, im, rot);
@@ -93,19 +93,19 @@ void RotatePointAroundVector(vec3_t dst, const vec3_t dir, const vec3_t point, f
 
 void AngleVectors(vec3_t angles, vec3_t forward, vec3_t right, vec3_t up)
 {
-	float		angle;
-	static float		sr, sp, sy, cr, cp, cy;
+	float angle;
+	static float sr, sp, sy, cr, cp, cy;
 	// static to help MS compiler fp bugs
 
 	angle = angles[YAW] * (M_PI * 2 / 360);
-	sy = sin(angle);
-	cy = cos(angle);
+	sy = sinf(angle);
+	cy = cosf(angle);
 	angle = angles[PITCH] * (M_PI * 2 / 360);
-	sp = sin(angle);
-	cp = cos(angle);
+	sp = sinf(angle);
+	cp = cosf(angle);
 	angle = angles[ROLL] * (M_PI * 2 / 360);
-	sr = sin(angle);
-	cr = cos(angle);
+	sr = sinf(angle);
+	cr = cosf(angle);
 
 	if (forward)
 	{
@@ -162,10 +162,10 @@ void PerpendicularVector(vec3_t dst, const vec3_t src)
 	*/
 	for (pos = 0, i = 0; i < 3; i++)
 	{
-		if (fabs(src[i]) < minelem)
+		if (fabsf(src[i]) < minelem)
 		{
 			pos = i;
-			minelem = fabs(src[i]);
+			minelem = fabsf(src[i]);
 		}
 	}
 	tempvec[0] = tempvec[1] = tempvec[2] = 0.0F;
@@ -272,9 +272,9 @@ float LerpAngle(float a2, float a1, float frac)
 }
 
 
-float	anglemod(float a)
+float anglemod(float a)
 {
-	a = (360.0 / 65536) * ((int32_t)(a * (65536 / 360.0)) & 65535);
+	a = (360.0f / 65536.0f) * ((int32_t)(a * (65536.0f / 360.0f)) & 65535);
 	return a;
 }
 
@@ -424,7 +424,7 @@ vec_t VectorNormalize(vec3_t v)
 	float	length, ilength;
 
 	length = v[0] * v[0] + v[1] * v[1] + v[2] * v[2];
-	length = sqrt(length);		// FIXME
+	length = sqrtf(length);		// FIXME
 
 	if (length)
 	{
@@ -440,10 +440,10 @@ vec_t VectorNormalize(vec3_t v)
 
 vec_t VectorNormalize2(vec3_t v, vec3_t out)
 {
-	float	length, ilength;
+	float length, ilength;
 
 	length = v[0] * v[0] + v[1] * v[1] + v[2] * v[2];
-	length = sqrt(length);		// FIXME
+	length = (float)sqrt(length);		// FIXME
 
 	if (length)
 	{
@@ -481,7 +481,7 @@ vec_t VectorLength(vec3_t v)
 	length = 0;
 	for (i = 0; i < 3; i++)
 		length += v[i] * v[i];
-	length = sqrt(length);		// FIXME
+	length = sqrtf(length);		// FIXME
 
 	return length;
 }
